Release 0.8.2
A fifth fuzzing round, run before 0.8.1 went to crates.io, found:

- CRW: sibling entries pointing at the same sub-directory re-read it at
  every level (54 KB file -> over 2 GB). Directories are read once, like
  ExifTool's PROCESSED check.
- PDF: octal string escapes above \377 overflowed a byte.

Regression inputs added. Parity audit: 196/196 identical.

+++ src/formats/pdf.rs
@@ -272,16 +272,17 @@ fn decode_pdf_literal_bytes(raw: &[u8]) -> String {
                 b')' => bytes.push(b')'),
                 b'\\' => bytes.push(b'\\'),
                 b'0'..=b'7' => {
-                    let mut val = raw[i] - b'0';
+                    let mut val = u16::from(raw[i] - b'0');
                     if i + 1 < raw.len() && raw[i + 1] >= b'0' && raw[i + 1] <= b'7' {
                         i += 1;
-                        val = val * 8 + (raw[i] - b'0');
+                        val = val * 8 + u16::from(raw[i] - b'0');
                         if i + 1 < raw.len() && raw[i + 1] >= b'0' && raw[i + 1] <= b'7' {
                             i += 1;
-                            val = val * 8 + (raw[i] - b'0');
+                            val = val * 8 + u16::from(raw[i] - b'0');
                         }
                     }
-                    bytes.push(val);
+                    // \400-\777 exceed a byte; the PDF spec drops the high bit.
+                    bytes.push(val as u8);
                 }
                 c => {
                     bytes.push(b'\\');
